Move logging files into src/, include/ folders
